Fix #512
microtime() inexplicably returns '0.01234 1456789012' rather than a sane value. We only care about the fractional part here as it has more entropy.

+++ Sources/Subs.php
@@ -4153,7 +4153,7 @@ function setupMenuContext()
  */
 function smf_seed_generator()
 {
-	updateSettings(array('rand_seed' => microtime() * 1000000));
+	updateSettings(array('rand_seed' => (float) microtime() * 1000000));
 }
 
 /**
